+++ src/modules/validation/fixtures/email-validation.component.fixture.html
@@ -1,11 +1,11 @@
 <div>
-    <input 
-        [(ngModel)]="emailValidator"
-        skyEmailValidation
-        #email="ngModel"/>
-    <div class="sky-error-label" *ngIf="email.errors && email.errors.skyEmail && (email.dirty || email.touched)">
-        <div [hidden]="!email.errors.skyEmail.invalid">
-            Please enter valid email with format joe@abc.com
-        </div>
+  <input 
+    [(ngModel)]="emailValidator" 
+    skyEmailValidation 
+    #email="ngModel" />
+  <div class="sky-error-label" *ngIf="email.errors && email.errors.skyEmail && (email.dirty || email.touched)">
+    <div [hidden]="!email.errors.skyEmail.invalid">
+      Please enter valid email with format joe@abc.com
     </div>
+  </div>
 </div>
