+++ src/modules/validation/fixtures/url-validation.component.fixture.html
@@ -1,11 +1,11 @@
 <div>
-    <input 
-        [(ngModel)]="urlValidator"
-        skyUrlValidation
-        #url="ngModel"/>
-    <div class="sky-error-label" *ngIf="url.errors && url.errors.skyUrl && (url.dirty || url.touched)">
-        <div [hidden]="!url.errors.skyUrl.invalid">
-            Please enter valid url with format http(s)://site.domain
-        </div>
+  <input
+    [(ngModel)]="urlValidator" 
+    skyUrlValidation 
+    #url="ngModel" />
+  <div class="sky-error-label" *ngIf="url.errors && url.errors.skyUrl && (url.dirty || url.touched)">
+    <div [hidden]="!url.errors.skyUrl.invalid">
+      Please enter valid url with format http(s)://site.domain
     </div>
+  </div>
 </div>
