+++ src/modules/validation/index.ts
@@ -1,3 +1,4 @@
 export { SkyEmailValidationDirective } from './email-validation.directive';
 export { SkyUrlValidationDirective } from './url-validation.directive';
 export { SkyValidationModule } from './validation.module';
+export { SkyValidator } from './validator';
