+++ src/modules/validation/url-validation.component.spec.ts
@@ -6,88 +6,88 @@ import { SkyUrlValidationFixturesModule } from './fixtures/url-validation-fixtur
 import { UrlValidationTestComponent } from './fixtures/url-validation.component.fixture';
 
 describe('Url validation', () => {
-    function setInput(
-        element: HTMLElement,
-        text: string,
-        compFixture: ComponentFixture<any>) {
-        let inputEvent = document.createEvent('Event');
-        let params = {
-            bubbles: false,
-            cancelable: false
-        };
+  function setInput(
+    element: HTMLElement,
+    text: string,
+    compFixture: ComponentFixture<any>) {
+    let inputEvent = document.createEvent('Event');
+    let params = {
+      bubbles: false,
+      cancelable: false
+    };
 
-        inputEvent.initEvent('input', params.bubbles, params.cancelable);
+    inputEvent.initEvent('input', params.bubbles, params.cancelable);
 
-        let changeEvent = document.createEvent('Event');
-        changeEvent.initEvent('change', params.bubbles, params.cancelable);
-        let inputEl = element.querySelector('input');
-        inputEl.value = text;
+    let changeEvent = document.createEvent('Event');
+    changeEvent.initEvent('change', params.bubbles, params.cancelable);
+    let inputEl = element.querySelector('input');
+    inputEl.value = text;
 
-        inputEl.dispatchEvent(inputEvent);
-        compFixture.detectChanges();
+    inputEl.dispatchEvent(inputEvent);
+    compFixture.detectChanges();
 
-        inputEl.dispatchEvent(changeEvent);
-        compFixture.detectChanges();
-        tick();
+    inputEl.dispatchEvent(changeEvent);
+    compFixture.detectChanges();
+    tick();
 
-    }
+  }
 
-    let component: UrlValidationTestComponent;
-    let fixture: ComponentFixture<UrlValidationTestComponent>;
-    let ngModel: NgModel;
-    let nativeElement: HTMLElement;
+  let component: UrlValidationTestComponent;
+  let fixture: ComponentFixture<UrlValidationTestComponent>;
+  let ngModel: NgModel;
+  let nativeElement: HTMLElement;
 
-    beforeEach(() => {
-        TestBed.configureTestingModule({
-            imports: [
-                SkyUrlValidationFixturesModule,
-                FormsModule
-            ]
-        });
-        fixture = TestBed.createComponent(UrlValidationTestComponent);
-        nativeElement = fixture.nativeElement as HTMLElement;
-        let input = fixture.debugElement.query(By.css('input'));
-        ngModel = <NgModel>input.injector.get(NgModel);
-        component = fixture.componentInstance;
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        SkyUrlValidationFixturesModule,
+        FormsModule
+      ]
     });
+    fixture = TestBed.createComponent(UrlValidationTestComponent);
+    nativeElement = fixture.nativeElement as HTMLElement;
+    let input = fixture.debugElement.query(By.css('input'));
+    ngModel = <NgModel>input.injector.get(NgModel);
+    component = fixture.componentInstance;
+  });
 
-    it('should validate correct input', fakeAsync(() => {
-        fixture.detectChanges();
-        tick();
-        setInput(nativeElement, 'https://blackbaud.com', fixture);
-        fixture.detectChanges();
+  it('should validate correct input', fakeAsync(() => {
+    fixture.detectChanges();
+    tick();
+    setInput(nativeElement, 'https://blackbaud.com', fixture);
+    fixture.detectChanges();
 
-        expect(nativeElement.querySelector('input').value).toBe('https://blackbaud.com');
+    expect(nativeElement.querySelector('input').value).toBe('https://blackbaud.com');
 
-        expect(ngModel.control.valid).toBe(true);
-        expect(ngModel.control.pristine).toBe(false);
-        expect(ngModel.control.touched).toBe(false);
+    expect(ngModel.control.valid).toBe(true);
+    expect(ngModel.control.pristine).toBe(false);
+    expect(ngModel.control.touched).toBe(false);
 
-    }));
-    it('should validate incorrect input', fakeAsync(() => {
-        fixture.detectChanges();
-        tick();
-        setInput(nativeElement, '[]awefhawenfc0293ejwf]', fixture);
-        fixture.detectChanges();
+  }));
+  it('should validate incorrect input', fakeAsync(() => {
+    fixture.detectChanges();
+    tick();
+    setInput(nativeElement, '[]awefhawenfc0293ejwf]', fixture);
+    fixture.detectChanges();
 
-        expect(nativeElement.querySelector('input').value).toBe('[]awefhawenfc0293ejwf]');
+    expect(nativeElement.querySelector('input').value).toBe('[]awefhawenfc0293ejwf]');
 
-        expect(ngModel.control.valid).toBe(false);
-        expect(ngModel.control.pristine).toBe(false);
-        expect(ngModel.control.touched).toBe(false);
-    }));
+    expect(ngModel.control.valid).toBe(false);
+    expect(ngModel.control.pristine).toBe(false);
+    expect(ngModel.control.touched).toBe(false);
+  }));
 
-    it('should handle invalid and then valid input', fakeAsync(() => {
-        fixture.detectChanges();
-        tick();
-        setInput(nativeElement, '[]awefhawenfc0293ejwf]', fixture);
-        setInput(nativeElement, 'blackbaud.com', fixture);
+  it('should handle invalid and then valid input', fakeAsync(() => {
+    fixture.detectChanges();
+    tick();
+    setInput(nativeElement, '[]awefhawenfc0293ejwf]', fixture);
+    setInput(nativeElement, 'blackbaud.com', fixture);
 
-        expect(nativeElement.querySelector('input').value).toBe('blackbaud.com');
-        expect(component.urlValidator).toEqual('blackbaud.com');
+    expect(nativeElement.querySelector('input').value).toBe('blackbaud.com');
+    expect(component.urlValidator).toEqual('blackbaud.com');
 
-        expect(ngModel.control.valid).toBe(true);
-        expect(ngModel.control.pristine).toBe(false);
-        expect(ngModel.control.touched).toBe(false);
-    }));
+    expect(ngModel.control.valid).toBe(true);
+    expect(ngModel.control.pristine).toBe(false);
+    expect(ngModel.control.touched).toBe(false);
+  }));
 });
