+++ src/modules/validation/url-validation.directive.ts
@@ -1,7 +1,7 @@
 import { Directive, forwardRef } from '@angular/core';
 import { NG_VALIDATORS } from '@angular/forms';
 
-import { SkyValidator } from './validation';
+import { SkyValidator } from './validator';
 
 // tslint:disable:no-forward-ref no-use-before-declare
 const SKY_URL_VALIDATION_VALIDATOR = {
